feat: enable batch manual errmsgs in client before uploading
this runs prudence schemas on the browser

straightforward enough, required a little refactoring, but we're good.
