+++ client/package.json
@@ -60,6 +60,7 @@
 		"postcss-safe-parser": "4.0.1",
 		"pretty-bytes": "5.6.0",
 		"prop-types": "15.6.0",
+		"prudence": "0.9.8",
 		"react": "17.0.2",
 		"react-async": "10.0.1",
 		"react-bootstrap": "1.6.1",
