+++ common/package.json
@@ -17,6 +17,7 @@
 	"dependencies": {
 		"@types/express-serve-static-core": "4.17.29",
 		"@types/mongodb": "3.6.18",
+		"deepmerge": "^4.2.2",
 		"prudence": "0.9.8"
 	},
 	"devDependencies": {
